// src/pages/UserManage/typings.d.ts
// @ts-ignore
/* eslint-disable */

declare namespace API {
  type Params = {
    pageNumber: number,
    pageSize: number,
    phone: string,
    realName: string
  };

  type TableItem = {
    address: string,
    birthday: string,
    emergencyContact: string,
    emergencyTel: string,
    idNumber: string,
    password: string,
    phone: string,
    realName: string,
    sex: number,
    userId: number,
    userName: string
  }
}